test(cli): contract signatures verification
